+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## [2.0.0] - 2026-03-11
+- Add function `createDebug`
+- Add method `dump`
+- Add option `native`
+- Add option `local`
+- Removed singleton
+- Target ES version (`ES2018`)
+
 ## [1.0.1] - 2026-01-06
 - Updated README.md